# docs/constellation-mcp-openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Constellation MCP",
    "description": "Model Context Protocol endpoints for Constellation consulting delivery platform. Read endpoints at /mcp/* provide tenant-scoped access to projects, assignments, time entries, expenses, RAIDD, financials, and CRM data. Write endpoints at /mcp/v1/* (Phase 0 — clients) are gated by the MCP_WRITES_ENABLED feature flag and require an X-Idempotency-Key header.",
    "version": "1.1.0"
  },
  "host": "constellation.synozur.com",
  "basePath": "/mcp",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/v1/clients": {
      "post": {
        "operationId": "CreateClient",
        "summary": "Create a client (with duplicate detection)",
        "description": "Creates a client in the caller's tenant. Performs a near-match scan against existing client names first — if any similar client is found, returns 409 with candidate records unless force=true is set in the body. Supports ?dryRun=true for conversational preview. Requires admin, pm, or portfolio-manager role.",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "Client-supplied UUID or opaque token. Replays with the same key return the cached response."
          },
          {
            "name": "dryRun",
            "in": "query",
            "type": "boolean",
            "required": false,
            "description": "If true, validates and returns the would-be client plus near-matches, but does not write to the database."
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name"],
              "properties": {
                "name": { "type": "string", "description": "Client display name (required)." },
                "shortName": { "type": "string", "description": "Abbreviated name for display (e.g. 'MSFT' for Microsoft)." },
                "currency": { "type": "string", "description": "ISO 4217 currency code (default 'USD')." },
                "status": { "type": "string", "enum": ["pending", "active", "inactive", "archived"], "description": "Initial status (default 'pending')." },
                "contactName": { "type": "string" },
                "billingContact": { "type": "string" },
                "force": { "type": "boolean", "description": "If true, bypasses near-match duplicate detection. Use only after confirming with the end user." }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Client created",
            "schema": {
              "type": "object",
              "properties": {
                "data": { "type": "object" },
                "created": { "type": "boolean" },
                "nearMatches": { "type": "array", "items": { "type": "object" } },
                "forced": { "type": "boolean" },
                "idempotent": { "type": "boolean" },
                "auditId": { "type": "string" },
                "correlationId": { "type": "string" }
              }
            }
          },
          "409": {
            "description": "Near-match to existing client — confirm with user and retry with force=true",
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "code": { "type": "string" },
                "candidates": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "name": { "type": "string" },
                      "status": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://login.microsoftonline.com",
      "tokenUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/token",
      "scopes": {
        "api://198aa0a6-d2ed-4f35-b41b-b6f6778a30d6/access_as_user": "Access Constellation MCP as signed-in user"
      }
    }
  },
  "security": [
    {
      "oauth2": ["api://198aa0a6-d2ed-4f35-b41b-b6f6778a30d6/access_as_user"]
    }
  ]
}